sanctumを使った認証周りの実装

+++ src/app/Http/Controllers/ApiTodoController.php
@@ -5,21 +5,42 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Todo;
 use App\Http\Requests\TodoRequest;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class ApiTodoController extends Controller
 {
     public function getAll()
     {
-        $todos = Todo::get();
-        
-        return response()->json($todos);
+        $userId = Auth::id();
+        $todos = Todo::where('user_id', $userId)->get();
+
+        if ($todos->isEmpty()) {
+          return response()->json([
+              'message' => 'todoが見つかりません'
+          ], 404);
+        }
+
+        return response()->json([
+            'todos' => $todos
+        ], 200);
 
     }
 
     public function create(Request $request)
     {
+        $userId = Auth::id();
         $todoInput = $request->all();
-        Todo::create($todoInput);
+        $name = $todoInput['name'];
+        $description = $todoInput['description'];
+        $category_id = $todoInput['category_id'];
+
+        Todo::create([
+          'name' => $name,
+          'description' => $description,
+          'user_id' => $userId,
+          'category_id' => $category_id,
+        ]);
 
         return response()->json($todoInput);
 
@@ -28,8 +49,20 @@ class ApiTodoController extends Controller
     public function updateById(Request $request, $id)
     {
         $todo = Todo::find($id);
-        $todo->update($request->all());
 
+        if (is_null($todo)) {
+          return response()->json([
+              'message' => 'todoが見つかりません'
+          ], 404);
+        }
+        $authUserId = Auth::id();
+        $userId = $todo['user_id'];
+        if ($authUserId !== $userId) {
+          return response()->json([
+              'message' => '認証に失敗しました'
+          ], 401);
+        }
+        $todo->update($request->all());
         return response()->json($request->all());
 
     }
@@ -37,16 +70,42 @@ class ApiTodoController extends Controller
     public function deleteById($id)
     {
         $todo = Todo::find($id);
-        $todo->delete();
 
+        if (is_null($todo)) {
+          return response()->json([
+              'message' => 'todoが見つかりません'
+          ], 404);
+        }
+        $authUserId = Auth::id();
+        $userId = $todo['user_id'];
+        if ($authUserId !== $userId) {
+          return response()->json([
+              'message' => '認証に失敗しました'
+          ], 401);
+        }
+        $todo->delete();
         return response()->json($todo);
 
     }
     public function getById(Request $request, $id)
     {
         $todo = Todo::find($id);
+        if (!$todo) {
+          return response()->json([
+              'message' => 'todoが見つかりません'
+          ], 404);
+        }
+        $authUserId = Auth::id();
+        $userId = $todo['user_id'];
+        if ($authUserId !== $userId) {
+          return response()->json([
+              'message' => '認証に失敗しました'
+          ], 401);
+        }
 
-        return response()->json($todo);
+        return response()->json([
+            'todo' => $todo
+        ], 200);
 
     }
 }
